apps/shipment-tracker/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "shipment-tracker"
version = "0.1.0"
description = "Charcoal shipment intelligence platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "psycopg[binary]>=3.1",
    "psycopg-pool>=3.1",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "httpx>=0.27",
    "loguru>=0.7",
    "websockets>=12.0",
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.3"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
